still need to complete some features Database

+++ app/src/main/java/com/chandan/android/movietowatch/model/Movie.java
@@ -128,7 +128,7 @@ public class Movie implements Parcelable {
         return voteAverage;
     }
 
-    public void setVoteAverage(double voteAverage) {
+    public void setVoteAverage(float voteAverage) {
         this.voteAverage = voteAverage;
     }
